rename TopNav elements
